[mod]因为底层结构体字段类型调整，所以同步修改了WTSDataFactory

+++ src/WTSTools/WTSDataFactory.cpp
@@ -90,13 +90,13 @@ WTSBarStruct* WTSDataFactory::updateMin1Data(WTSSessionInfo* sInfo, WTSKlineData
 	uint32_t uMinute = sInfo->timeToMinutes(uTime) - 1;
 
 	uint32_t uBarMin = (uMinute / steplen)*steplen + steplen;
-	uint32_t uBarTime = sInfo->minuteToTime(uBarMin);
+	uint64_t uBarTime = sInfo->minuteToTime(uBarMin);
 	//if(uBarTime > uTime && !sInfo->isInAuctionTime(uTime))
 	//{
 	//	//这种情况只可能是日期倒退
 	//	uDate = TimeUtils::getNextDate(uDate, -1);
 	//}
-	uBarTime = TimeUtils::timeToMinBar(uDate, uBarTime);
+	uBarTime = TimeUtils::timeToMinBar(uDate, (uint32_t)uBarTime);
 
 	WTSBarStruct* lastBar = NULL;
 	if (klineData->size() > 0)
@@ -185,9 +185,9 @@ WTSBarStruct* WTSDataFactory::updateMin1Data(WTSSessionInfo* sInfo, WTSKlineData
 	{
 		uDate = TimeUtils::getNextDate(uDate);
 	}
-	uint32_t uBarTime = TimeUtils::timeToMinBar(uDate, uOnlyMin);
+	uint64_t uBarTime = TimeUtils::timeToMinBar(uDate, uOnlyMin);
 
-	uint32_t lastTime = klineData->time(-1);
+	uint64_t lastTime = klineData->time(-1);
 	uint32_t lastDate = klineData->date(-1);
 	if (lastTime == INVALID_UINT32 || uBarTime > lastTime || tick->tradingdate() > lastDate)
 	{
@@ -251,8 +251,8 @@ WTSBarStruct* WTSDataFactory::updateMin5Data(WTSSessionInfo* sInfo, WTSKlineData
 
 
 	uint32_t uBarMin = (uMinute / steplen)*steplen + steplen;
-	uint32_t uBarTime = sInfo->minuteToTime(uBarMin);
-	uBarTime = TimeUtils::timeToMinBar(uDate, uBarTime);
+	uint64_t uBarTime = sInfo->minuteToTime(uBarMin);
+	uBarTime = TimeUtils::timeToMinBar(uDate, (uint32_t)uBarTime);
 
 	WTSBarStruct* lastBar = NULL;
 	if (klineData->size() > 0)
@@ -324,9 +324,9 @@ WTSBarStruct* WTSDataFactory::updateMin5Data(WTSSessionInfo* sInfo, WTSKlineData
 	{
 		uDate = TimeUtils::getNextDate(uDate);
 	}
-	uint32_t uBarTime = TimeUtils::timeToMinBar(uDate, uOnlyMin);
+	uint64_t uBarTime = TimeUtils::timeToMinBar(uDate, uOnlyMin);
 
-	uint32_t lastTime = klineData->time(klineData->size()-1);
+	uint64_t lastTime = klineData->time(klineData->size()-1);
 	if(lastTime == INVALID_UINT32 || uBarTime != lastTime)
 	{
 		//如果时间不一致,则新增一条K线
@@ -411,7 +411,7 @@ WTSBarStruct* WTSDataFactory::updateSecData(WTSSessionInfo* sInfo, WTSKlineData*
 		barTime = (uint32_t)(TimeUtils::makeTime(uDate, barTime * 1000) / 1000);
 	}	
 
-	uint32_t lastTime = klineData->time(klineData->size()-1);
+	uint64_t lastTime = klineData->time(klineData->size()-1);
 	if(lastTime == INVALID_UINT32 || lastTime != barTime)
 	{
 		WTSBarStruct *day = new WTSBarStruct;
@@ -514,13 +514,13 @@ WTSKlineData* WTSDataFactory::extractMin1Data(WTSKlineSlice* baseKline, uint32_t
 		uint32_t uMinute = sInfo->timeToMinutes(uTime)-1;
 
 		uint32_t uBarMin = (uMinute/steplen)*steplen + steplen;
-		uint32_t uBarTime = sInfo->minuteToTime(uBarMin);
+		uint64_t uBarTime = sInfo->minuteToTime(uBarMin);
 		//if(uBarTime > uTime && !sInfo->isInAuctionTime(uTime))
 		//{
 		//	//这种情况只可能是日期倒退
 		//	uDate = TimeUtils::getNextDate(uDate, -1);
 		//}
-		uBarTime = TimeUtils::timeToMinBar(uDate, uBarTime);
+		uBarTime = TimeUtils::timeToMinBar(uDate, (uint32_t)uBarTime);
 
 		WTSBarStruct* lastBar = NULL;
 		if(ret->size() > 0)
@@ -611,8 +611,8 @@ WTSKlineData* WTSDataFactory::extractMin5Data(WTSKlineSlice* baseKline, uint32_t
 
 
 		uint32_t uBarMin = (uMinute/steplen)*steplen+steplen;
-		uint32_t uBarTime = sInfo->minuteToTime(uBarMin);
-		uBarTime = TimeUtils::timeToMinBar(uDate, uBarTime);
+		uint64_t uBarTime = sInfo->minuteToTime(uBarMin);
+		uBarTime = TimeUtils::timeToMinBar(uDate, (uint32_t)uBarTime);
 
 		WTSBarStruct* lastBar = NULL;
 		if(ret->size() > 0)
@@ -837,7 +837,7 @@ bool WTSDataFactory::mergeKlineData(WTSKlineData* klineData, WTSKlineData* newKl
 	}
 	else
 	{
-		uint32_t sTime, eTime;
+		uint64_t sTime,eTime;
 		if(klineData->period() == KP_DAY)
 		{
 			sTime = bars[0].date;
@@ -855,7 +855,7 @@ bool WTSDataFactory::mergeKlineData(WTSKlineData* klineData, WTSKlineData* newKl
 		{
 			WTSBarStruct& curBar = newBars[i];
 
-			uint32_t curTime;
+			uint64_t curTime;
 			if (klineData->period() == KP_DAY)
 				curTime = curBar.date;
 			else
